Add searchable combobox to core.gui
