+++ QtBrainChartGUI/core/gui/SearchableQComboBox.py
@@ -1,0 +1,53 @@
+# This Python file uses the following encoding: utf-8
+"""
+Author: Ashish Singh
+contact: software@cbica.upenn.edu
+Copyright (c) 2018 University of Pennsylvania. All rights reserved.
+Use of this source code is governed by license located in license file: https://github.com/CBICA/BrainChart/blob/main/LICENSE
+"""
+from PyQt5 import QtCore, QtWidgets
+
+class SearchableQComboBox(QtWidgets.QComboBox):
+    def __init__(self, parent=None):
+        super(SearchableQComboBox, self).__init__(parent)
+
+        self.setFocusPolicy(QtCore.Qt.StrongFocus)
+        self.setEditable(True)
+
+        # add a filter model to filter matching items
+        self.pFilterModel = QtCore.QSortFilterProxyModel(self)
+        self.pFilterModel.setFilterCaseSensitivity(QtCore.Qt.CaseInsensitive)
+        self.pFilterModel.setSourceModel(self.model())
+
+        # add a completer, which uses the filter model
+        self.completer = QtWidgets.QCompleter(self.pFilterModel, self)
+        # always show all (filtered) completions
+        self.completer.setCompletionMode(QtWidgets.QCompleter.UnfilteredPopupCompletion)
+        self.setCompleter(self.completer)
+
+        # connect signals
+        self.lineEdit().textEdited.connect(self.pFilterModel.setFilterFixedString)
+        self.completer.activated.connect(self.on_completer_activated)
+
+
+    # on selection of an item from the completer, select the corresponding item from combobox
+    def on_completer_activated(self, text):
+        if text:
+            index = self.findText(text)
+            self.setCurrentIndex(index)
+            self.activated[str].emit(self.itemText(index))
+
+
+    # on model change, update the models of the filter and completer as well
+    def setModel(self, model):
+        super(SearchableQComboBox, self).setModel(model)
+        self.pFilterModel.setSourceModel(model)
+        self.completer.setModel(self.pFilterModel)
+
+
+    # on model column change, update the model column of the filter and completer as well
+    def setModelColumn(self, column):
+        self.completer.setCompletionColumn(column)
+        self.pFilterModel.setFilterKeyColumn(column)
+        super(SearchableQComboBox, self).setModelColumn(column)
+
