+++ QtBrainChartGUI/plugins/harmonization/harmonization.py
@@ -8,50 +8,7 @@ import numpy as np
 import pandas as pd
 from QtBrainChartGUI.core.plotcanvas import PlotCanvas
 from QtBrainChartGUI.core.baseplugin import BasePlugin
-
-class ExtendedComboBox(QtWidgets.QComboBox):
-    def __init__(self, parent=None):
-        super(ExtendedComboBox, self).__init__(parent)
-
-        self.setFocusPolicy(QtCore.Qt.StrongFocus)
-        self.setEditable(True)
-
-        # add a filter model to filter matching items
-        self.pFilterModel = QtCore.QSortFilterProxyModel(self)
-        self.pFilterModel.setFilterCaseSensitivity(QtCore.Qt.CaseInsensitive)
-        self.pFilterModel.setSourceModel(self.model())
-
-        # add a completer, which uses the filter model
-        self.completer = QtWidgets.QCompleter(self.pFilterModel, self)
-        # always show all (filtered) completions
-        self.completer.setCompletionMode(QtWidgets.QCompleter.UnfilteredPopupCompletion)
-        self.setCompleter(self.completer)
-
-        # connect signals
-        self.lineEdit().textEdited.connect(self.pFilterModel.setFilterFixedString)
-        self.completer.activated.connect(self.on_completer_activated)
-
-
-    # on selection of an item from the completer, select the corresponding item from combobox 
-    def on_completer_activated(self, text):
-        if text:
-            index = self.findText(text)
-            self.setCurrentIndex(index)
-            self.activated[str].emit(self.itemText(index))
-
-
-    # on model change, update the models of the filter and completer as well 
-    def setModel(self, model):
-        super(ExtendedComboBox, self).setModel(model)
-        self.pFilterModel.setSourceModel(model)
-        self.completer.setModel(self.pFilterModel)
-
-
-    # on model column change, update the model column of the filter and completer as well
-    def setModelColumn(self, column):
-        self.completer.setCompletionColumn(column)
-        self.pFilterModel.setFilterKeyColumn(column)
-        super(ExtendedComboBox, self).setModelColumn(column)  
+from QtBrainChartGUI.core.gui.SearchableQComboBox import SearchableQComboBox
 
 class Harmonization(QtWidgets.QWidget,BasePlugin):
 
@@ -63,7 +20,7 @@ class Harmonization(QtWidgets.QWidget,BasePlugin):
         self.readAdditionalInformation(root)
         self.ui = uic.loadUi(os.path.join(root, 'harmonization.ui'),self)
         self.ui.Harmonization_Model_Loaded_Lbl.setHidden(True)
-        self.ui.comboBoxROI = ExtendedComboBox(self.ui)
+        self.ui.comboBoxROI = SearchableQComboBox(self.ui)
         self.plotCanvas = PlotCanvas(self.ui.page_2)
         self.plotCanvas.axes1 = self.plotCanvas.fig.add_subplot(131)
         self.plotCanvas.axes2 = self.plotCanvas.fig.add_subplot(132)
